+++ lib/utils.py
@@ -1,0 +1,6 @@
+import random
+
+# FUNCTIONS
+def gen_charid():
+  int = random.randint(100000, 999999)
+  return int