+++ smb/admin.py
@@ -1,0 +1,6 @@
+from django.contrib import admin
+
+from .models import Mech, MechClass
+
+admin.site.register(Mech)
+admin.site.register(MechClass)