+++ smb/rules.py
@@ -1,0 +1,8 @@
+ability_level = [
+  {'score': [0, 1], 'die': '1d4'},
+  {'score': [2, 3], 'die': '1d6'},
+  {'score': [4, 5], 'die': '1d8'},
+  {'score': [6, 7], 'die': '1d10'},
+  {'score': [8, 9], 'die': '1d12'},
+  {'score': [10], 'die': '1d20'},
+]